requirements.txt
image-similarity-measures == 0.3.6
numpy == 1.24.4
opencv-python == 4.8.0.74
pandas == 2.0.3
pillow-avif-plugin == 1.3.1
pyfftw == 0.13.1
streamlit == 1.24.0